Declare conscrypt's min_sdk_version in Android.bp
`apexer` is used to build *.apex files. There are two inputs to `apexer`
which determine the minSdkVersion of the final .apex artifact
1. AndroidManifest.xml
2. min_sdk_version declared in Android.bp
(in that order)

Move the min_sdk_version of conscrypt from (1) to (2). The use case for
this is to build the AOSP and Google conscrypt apexes with
min_sdk_version of 30, but Google go conscrypt apexes with
min_sdk_version 33 (Go apexes were introduced in 33)

Bug: 295311875
Test: m com.android.conscrypt com.google.androd.conscrypt
com.google.android.go.conscrypt (with ag/27464797)
Test: aapt2 dump badging <path_to_conscrypt_apex(es)>
Test: verified that aosp and google BA apex have min_sdk_version of 30,
while google go apex has min_sdk_version of 33

Change-Id: Ia0d28cc9c903bf2a370ebc8f66c330e09f6bff8c